backend/requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
Werkzeug==2.3.8
spacy==3.7.1
pdfplumber==0.8.0
python-docx==0.8.11
python-dotenv==1.0.0
